+++ dart/geobase/lib/src/vector_data/model/geometry/multi_linestring.dart
@@ -172,6 +172,7 @@ class MultiLineString extends SimpleGeometry {
   ///
   /// Use [crs] to give hints (like axis order, and whether x and y must
   /// be swapped when read in) about coordinate reference system in text input.
+  /// When data itself have CRS information it overrides this value.
   ///
   /// Format or decoder implementation specific options can be set by [options].
   ///
@@ -334,6 +335,10 @@ class MultiLineString extends SimpleGeometry {
   /// When [format] is not given, then the geometry format of [WKB] is used as
   /// a default.
   ///
+  /// Use [crs] to give hints (like axis order, and whether x and y must
+  /// be swapped when read in) about coordinate reference system in binary
+  /// input. When data itself have CRS information it overrides this value.
+  ///
   /// Format or decoder implementation specific options can be set by [options].
   ///
   /// See also [MultiLineString.decodeHex] to decode from bytes represented as
@@ -341,11 +346,13 @@ class MultiLineString extends SimpleGeometry {
   factory MultiLineString.decode(
     Uint8List bytes, {
     BinaryFormat<SimpleGeometryContent> format = WKB.geometry,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) =>
       GeometryBuilder.decode<MultiLineString>(
         bytes,
         format: format,
+        crs: crs,
         options: options,
       );
 
@@ -364,11 +371,13 @@ class MultiLineString extends SimpleGeometry {
   factory MultiLineString.decodeHex(
     String bytesHex, {
     BinaryFormat<SimpleGeometryContent> format = WKB.geometry,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) =>
       GeometryBuilder.decodeHex<MultiLineString>(
         bytesHex,
         format: format,
+        crs: crs,
         options: options,
       );