+++ dart/geobase/lib/src/vector_data/model/geometry/multi_polygon.dart
@@ -238,6 +238,7 @@ class MultiPolygon extends SimpleGeometry {
   ///
   /// Use [crs] to give hints (like axis order, and whether x and y must
   /// be swapped when read in) about coordinate reference system in text input.
+  /// When data itself have CRS information it overrides this value.
   ///
   /// Format or decoder implementation specific options can be set by [options].
   ///
@@ -466,6 +467,10 @@ class MultiPolygon extends SimpleGeometry {
   /// When [format] is not given, then the geometry format of [WKB] is used as
   /// a default.
   ///
+  /// Use [crs] to give hints (like axis order, and whether x and y must
+  /// be swapped when read in) about coordinate reference system in binary
+  /// input. When data itself have CRS information it overrides this value.
+  ///
   /// Format or decoder implementation specific options can be set by [options].
   ///
   /// See also [MultiPolygon.decodeHex] to decode from bytes represented as a
@@ -473,11 +478,13 @@ class MultiPolygon extends SimpleGeometry {
   factory MultiPolygon.decode(
     Uint8List bytes, {
     BinaryFormat<SimpleGeometryContent> format = WKB.geometry,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) =>
       GeometryBuilder.decode<MultiPolygon>(
         bytes,
         format: format,
+        crs: crs,
         options: options,
       );
 
@@ -496,11 +503,13 @@ class MultiPolygon extends SimpleGeometry {
   factory MultiPolygon.decodeHex(
     String bytesHex, {
     BinaryFormat<SimpleGeometryContent> format = WKB.geometry,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) =>
       GeometryBuilder.decodeHex<MultiPolygon>(
         bytesHex,
         format: format,
+        crs: crs,
         options: options,
       );