+++ dart/geobase/lib/src/vector_data/model/geometry/polygon.dart
@@ -367,6 +367,7 @@ class Polygon extends SimpleGeometry {
   ///
   /// Use [crs] to give hints (like axis order, and whether x and y must
   /// be swapped when read in) about coordinate reference system in text input.
+  /// When data itself have CRS information it overrides this value.
   ///
   /// Format or decoder implementation specific options can be set by [options].
   ///
@@ -652,6 +653,10 @@ class Polygon extends SimpleGeometry {
   /// When [format] is not given, then the geometry format of [WKB] is used as
   /// a default.
   ///
+  /// Use [crs] to give hints (like axis order, and whether x and y must
+  /// be swapped when read in) about coordinate reference system in binary
+  /// input. When data itself have CRS information it overrides this value.
+  ///
   /// Format or decoder implementation specific options can be set by [options].
   ///
   /// See also [Polygon.decodeHex] to decode from bytes represented as a hex
@@ -659,11 +664,13 @@ class Polygon extends SimpleGeometry {
   factory Polygon.decode(
     Uint8List bytes, {
     BinaryFormat<SimpleGeometryContent> format = WKB.geometry,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) =>
       GeometryBuilder.decode<Polygon>(
         bytes,
         format: format,
+        crs: crs,
         options: options,
       );
 
@@ -682,11 +689,13 @@ class Polygon extends SimpleGeometry {
   factory Polygon.decodeHex(
     String bytesHex, {
     BinaryFormat<SimpleGeometryContent> format = WKB.geometry,
+    CoordRefSys? crs,
     Map<String, dynamic>? options,
   }) =>
       GeometryBuilder.decodeHex<Polygon>(
         bytesHex,
         format: format,
+        crs: crs,
         options: options,
       );